data_collectors/requirements.txt
# Data Collection Service Dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
aiohttp>=3.9.0
asyncio-throttle>=1.0.2
schedule>=1.2.0
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic>=2.5.0
requests>=2.31.0
sqlalchemy>=2.0.23
alembic>=1.13.0

# Web scraping dependencies
beautifulsoup4>=4.12.2
lxml>=4.9.3
selenium>=4.15.0

# Data processing
pandas>=2.1.3
numpy>=1.25.2

# Logging and monitoring
structlog>=23.2.0
